+++ crates/nu-command/src/default_context.rs
@@ -187,12 +187,12 @@ pub fn create_default_context(cwd: impl AsRef<Path>) -> EngineState {
             AnsiGradient,
             AnsiStrip,
             Clear,
-            DefaultKeybindings,
+            KeybindingsDefault,
             Input,
-            InputKeys,
+            KeybindingsListen,
             Keybindings,
             Kill,
-            ListKeybindings,
+            KeybindingsList,
             Sleep,
             TermSize,
         };
