+++ crates/nu-command/src/filters/keep/keep_while.rs
@@ -7,11 +7,11 @@ use nu_protocol::{
 };
 
 #[derive(Clone)]
-pub struct SkipUntil;
+pub struct KeepWhile;
 
-impl Command for SkipUntil {
+impl Command for KeepWhile {
     fn name(&self) -> &str {
-        "skip until"
+        "keep while"
     }
 
     fn signature(&self) -> Signature {
@@ -19,21 +19,21 @@ impl Command for SkipUntil {
             .required(
                 "predicate",
                 SyntaxShape::RowCondition,
-                "the predicate that skipped element must not match",
+                "the predicate that kept element must not match",
             )
             .category(Category::Filters)
     }
 
     fn usage(&self) -> &str {
-        "Skip elements of the input until a predicate is true."
+        "Keep elements of the input while a predicate is true."
     }
 
     fn examples(&self) -> Vec<Example> {
         vec![Example {
-            description: "Skip until the element is positive",
-            example: "echo [-2 0 2 -1] | skip until $it > 0",
+            description: "Keep while the element is negative",
+            example: "echo [-1 -2 9 1] | keep while $it < 0",
             result: Some(Value::List {
-                vals: vec![Value::test_int(2), Value::test_int(-1)],
+                vals: vec![Value::test_int(-1), Value::test_int(-2)],
                 span: Span::test_data(),
             }),
         }]
@@ -52,6 +52,7 @@ impl Command for SkipUntil {
 
         let block = engine_state.get_block(capture_block.block_id).clone();
         let var_id = block.signature.get_positional(0).and_then(|arg| arg.var_id);
+
         let mut stack = stack.captures_to_stack(&capture_block.captures);
 
         let ctrlc = engine_state.ctrlc.clone();
@@ -59,12 +60,12 @@ impl Command for SkipUntil {
 
         Ok(input
             .into_iter()
-            .skip_while(move |value| {
+            .take_while(move |value| {
                 if let Some(var_id) = var_id {
                     stack.add_var(var_id, value.clone());
                 }
 
-                !eval_block(&engine_state, &mut stack, &block, PipelineData::new(span))
+                eval_block(&engine_state, &mut stack, &block, PipelineData::new(span))
                     .map_or(false, |pipeline_data| {
                         pipeline_data.into_value(span).is_true()
                     })
@@ -75,12 +76,12 @@ impl Command for SkipUntil {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use keep_while::*;
 
     #[test]
     fn test_examples() {
         use crate::test_examples;
 
-        test_examples(SkipUntil)
+        test_examples(KeepWhile)
     }
 }
