+++ crates/nu-command/src/hash/hash.rs
