+++ crates/nu-command/src/math/math.rs
