+++ crates/nu-command/src/math/mod.rs
@@ -1,9 +1,9 @@
 mod abs;
 mod avg;
 mod ceil;
-pub mod command;
 mod eval;
 mod floor;
+pub mod math;
 mod max;
 mod median;
 mod min;
@@ -20,9 +20,9 @@ mod variance;
 pub use abs::SubCommand as MathAbs;
 pub use avg::SubCommand as MathAvg;
 pub use ceil::SubCommand as MathCeil;
-pub use command::MathCommand as Math;
 pub use eval::SubCommand as MathEval;
 pub use floor::SubCommand as MathFloor;
+pub use math::MathCommand as Math;
 pub use max::SubCommand as MathMax;
 pub use median::SubCommand as MathMedian;
 pub use min::SubCommand as MathMin;
