+++ crates/nu-command/src/platform/ansi/mod.rs
@@ -1,7 +1,7 @@
-mod command;
+mod ansi;
 mod gradient;
 mod strip;
 
-pub use command::AnsiCommand as Ansi;
+pub use ansi::AnsiCommand as Ansi;
 pub use gradient::SubCommand as AnsiGradient;
 pub use strip::SubCommand as AnsiStrip;
