+++ crates/nu-command/src/platform/reedline_commands/keybindings.rs
