+++ crates/nu-command/src/random/mod.rs
@@ -1,9 +1,9 @@
 mod bool;
 mod chars;
-mod command;
 mod decimal;
 mod dice;
 mod integer;
+mod random;
 mod uuid;
 
 pub use self::bool::SubCommand as RandomBool;
@@ -12,4 +12,4 @@ pub use self::decimal::SubCommand as RandomDecimal;
 pub use self::dice::SubCommand as RandomDice;
 pub use self::integer::SubCommand as RandomInteger;
 pub use self::uuid::SubCommand as RandomUuid;
-pub use command::RandomCommand as Random;
+pub use random::RandomCommand as Random;
