+++ crates/nu-command/src/random/random.rs
